Add files via upload
open Shopify code editor and then go for sections, In the section area, you will have to create a new section with "logo-slider" then select everything in that folder and replace it with provided code and save the section.
After doing the above task you will have to go back to the theme area and click on customize. Then click on add a new section and you will see "Logo Slider" there add it and enjoy.